// utils/constants.d.ts
export declare const COLLECTION_USER_DATA: string;
export declare const LINK_ICONS_METADATA: {
    label: string;
    value: string;
    icon: import("@fortawesome/fontawesome-common-types").IconDefinition;
}[];
export declare const linkIcons: import("@fortawesome/fontawesome-common-types").IconDefinition[];
export declare const KEYBOARD_KEY: {
    ENTER_KEY: string;
    ESCAPE_KEY: string;
    SPACE_KEY: string;
    LEFT_ARROW_KEY: string;
    UP_ARROW_KEY: string;
    RIGHT_ARROW_KEY: string;
    DOWN_ARROW_KEY: string;
};
export declare const imageTypes: string[];